run-vmtest: misc scripts updates
* remove boot parameter processing in run-bpf-selftests.sh
* add logging to run.sh re VMLINUX and VMLINUZ variables

+++ run-vmtest/run.sh
@@ -7,19 +7,25 @@ source "${GITHUB_ACTION_PATH}/../helpers.sh"
 
 export VMLINUZ=${VMLINUZ:-}
 if [[ ! -f "${VMLINUZ}" ]]; then
+    echo "Could not find VMLINUZ=\"$VMLINUZ\", searching with make -s image_name"
     image_name=$(make -C ${KERNEL_ROOT} -s image_name)
     export VMLINUZ=$(realpath ${KBUILD_OUTPUT})/${image_name}
 fi
 
+if [[ ! -f "${VMLINUZ}" ]]; then
+    echo "Could not find VMLINUZ (compressed kernel binary), exiting"
+    exit 2
+fi
+
 # Create a symlink to vmlinux from a "standard" location
 # See btf__load_vmlinux_btf() in libbpf
-VMLINUX=${VMLINUX:-"$KBUILD_OUTPUT/vmlinux"}
+export VMLINUX=${VMLINUX:-"$KBUILD_OUTPUT/vmlinux"}
 if [[ -f "${VMLINUX}" ]]; then
     VMLINUX_VERSION="$(strings ${VMLINUX} | grep -m 1 'Linux version' | awk '{print $3}')" || true
     sudo mkdir -p /usr/lib/debug/boot
     sudo ln -sf "${VMLINUX}" "/usr/lib/debug/boot/vmlinux-${VMLINUX_VERSION}"
 else
-    echo "$VMLINUX does not exist"
+    echo "Could not find VMLINUX=\"$VMLINUX\", exiting"
     exit 2
 fi
 
@@ -41,7 +47,7 @@ else
 fi
 
 # clear exitstatus file
-echo -n "" > exitstatus
+echo -n > exitstatus
 
 foldable start bpftool_checks "Running bpftool checks..."
